project/requirements/dev.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
crispy-bootstrap5==2024.2
Django==5.1
django-celery-results==2.5.1
django-crispy-forms==2.3
django-environ==0.11.2
django-ranged-response==0.2.0
django-redis==5.4.0
django-simple-captcha==0.6.0
gunicorn==23.0.0
kombu==5.4.0
packaging==24.1
pillow==10.4.0
prompt_toolkit==3.0.47
python-dateutil==2.9.0.post0
redis==5.0.8
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
